refactor(measurement): split sweep contracts and compiler
